mpharma/wsgi.py
"""
WSGI config for mpharma project.

It exposes the WSGI callable as a module-level variable named ``application``.

For more information on this file, see
https://docs.djangoproject.com/en/3.1/howto/deployment/wsgi/
"""

import os

from django.core.wsgi import get_wsgi_application
from dotenv import load_dotenv

# project_folder = os.path.expanduser('/home/ebo/src/mpharma/mpharma/')  # adjust as appropriate
# print (os.path.join(os.path.dirname(os.path.dirname(__file__)), '.env'))
load_dotenv(os.path.join(os.path.dirname(os.path.dirname(__file__)), '.env'))

os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'mpharma.settings')

application = get_wsgi_application()
