requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
gunicorn
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=1_gnu
asgiref=3.3.1=pypi_0
ca-certificates=2020.12.5=ha878542_0
certifi=2020.12.5=py39hf3d152e_1
django=3.1.5=pypi_0
djangorestframework=3.12.2=pypi_0
ld_impl_linux-64=2.35.1=hea4e1c9_1
libffi=3.3=h58526e2_2
libgcc-ng=9.3.0=h2828fa1_18
libgomp=9.3.0=h2828fa1_18
libstdcxx-ng=9.3.0=h6de172a_18
ncurses=6.2=h58526e2_4
openssl=1.1.1i=h7f98852_0
pip=20.3.3=pyhd8ed1ab_0
psycopg2=2.8.6=pypi_0
python=3.9.1=hffdb5ce_3_cpython
python-dotenv=0.15.0=pypi_0
python_abi=3.9=1_cp39
pytz=2020.5=pypi_0
readline=8.0=he28a2e2_2
setuptools=49.6.0=py39hf3d152e_3
sqlite=3.34.0=h74cdb3f_0
sqlparse=0.4.1=pypi_0
tk=8.6.10=h21135ba_1
tzdata=2020f=he74cb21_0
wheel=0.36.2=pyhd3deb0d_0
xz=5.2.5=h516909a_1
zlib=1.2.11=h516909a_1010
